[#25] engine: fix serialization of helper options functions to ruby
This fixes an issue where `mini_racer` does not support serializing javascript
functions nested in objects, such as in the options hash of helper function.

To avoid this issue, redefine any functions with a generic `"function"` string.

TODO: In the future, we can build a method to call these functions.